🎉 feat(delete-supplier.ts): add DeleteSupplierUseCase class to handle deleting a supplier
This commit adds a new file `delete-supplier.ts` which contains the `DeleteSupplierUseCase` class. This class is responsible for handling the use case of deleting a supplier.

The `DeleteSupplierUseCase` class takes a `suppliersRepository` as a dependency and has an `execute` method that takes a `supplierId` as a parameter. It first checks if the supplier exists in the repository by calling the `findById` method. If the supplier does not exist, it returns a `ResourceNotFoundError` using the `left` function from the `Either` type. If the supplier exists, it calls the `delete` method on the repository to delete the supplier.

The `execute` method returns a `DeleteSupplierUseCaseResponse` which is an `Either` type that can either be a `ResourceNotFoundError` or `null`.